requirements.txt
bitcoinlib
requests
mysql-connector-python
python-dotenv